:hammer: Make FileInfoTree creation multiplatform compatible
